Fix hunspell dicts detection
`hunspell -D` output is really different from `aspell dicts`.  It
outputs (1) a search path and (2) a list of paths to dictionary files.

(1) use a sed call to clean `hunspell -D` output
(2) change the filter to match at the end, whatever be on the beginning
of the line.

`hunspell -d <dict>` supports having a full path in argument, so we can
keep the full path.