Make `sed` call optional.
`sed` is not available e.g. on Windows.  If we do not have such a
command, don't use it, and just hope search path and section headers
lines won't be recognized as a usable dictionnary…

This should be OK most of the time: these lines should not end by $LANG.